feat: domain 패키지 생성
